AX_ENABLE_BUILDDIR: Renamed variable HOST to BUILD.
The patch universally replaces "host" with "build" and "HOST"
with "BUILD". The rationale is that typically the user wishes to
segregate builds based upon the BUILD target rather than the
configuration HOST type. Now that these host and build variables
are treated as more fully distinct in Autoconf, it makes sense to
honor this distinction.