AX_F90_MODULE_FLAG: support -module PGI compiler flag
I have "-module " to the list of compiler options that is tried
in order to support the PGI Fortran compiler. Without this
change, the PGI compiler actually accepts "-I" and passes the
test. However, this option is actually meant for setting the
include file search path to find files that are included with the
Fortran INCLUDE statement. Although the -I flag works in this
case because there are no -module options on the command line,
things will not work correctly if the user does bring in the
-module option on the actual compile command line.