AX_F90_MODULE_FLAG: drop support for -module flag
It turns out that in Version 3.0 of the Pathscale Fortran
compiler they have "clarified" the meaning of the -module flag to
only be for setting the directory where *.mod files are written.
In version 3.0, you can only use -module once in a command line,
and if you use it, then the compiler will look there first for
module files to be read before looking in the -I directories.
Thus, it really is best to just use the -I flag here and let the
user add one specific -module option only if they want or need to
redirect module output files.

This commit reverts 5ba234dd8f69456238e7718d4f3a7c116639e373.