AX_ENABLE_BUILDDIR: Revert renaming of variable HOST to BUILD.
Further discussion on the Autoconf mailing list has reached the
following consensus: although BUILD is the right choice in most
cases, in case of cross-compilation it is not. Cross-compilation
needs HOST to be set, so the original version of this macro is
the safer choice. This patch reverts commit
41da8f494cd209a9836ebeae8f211cb84dde9153.

An unpleasant property of HOST is that most shell environments
define that variable to the local hostname. When configure uses
$HOST as a default setting, however, that is not what it expects
to find -- HOST is supposed to be config.guess tripple. Thus, the
user is forced to specify HOST explicitly or to undefine $HOST in
the shell before calling configure. That's a bit of mess. Bright
ideas and patches are welcome.