AX_MISSING_PROG, AX_PATH_MISSING: re-factored for improved modularity
Added a new macro AX_MISSING_PROG that wraps AM_MISSING_PROG (from
Automake) and modified AX_PATH_MISSING to wrap AX_MISSING_PROG. The
change is transparent for users of the previous version. 

The names used in the macro-chain are confusing, but I didn't find a
better solution. Suggestions are welcome.