AX_ENABLE_BUILDDIR: Fix ":=" assignments in top-level generated Makefile
The sed script which generates the top-level Makefile from the Makefile
in the build directory is flawed.  It treats an assignment of the form
"TESTSUITE := ..." as a makefile rule - because it contains ":".

$ make
...
MAKE  : 1 * TESTSUITE
make[1]: Entering directory `/home/alan/bootup/src/module-init-tools/build'
make[1]: *** No rule to make target `TESTSUITE'.  Stop.

Fix the sed script to skip the rule-related commands for lines which
contain ":=".  As a cleanup, we can also skip the rule-related commands
for lines which do not contain ":".  That simplifies those commands,
because they don't need to test for ":" themselves.