AX_ENABLE_BUILDDIR: Fix path for config.guess
Using this macro, I get a warning that config.guess can't be found.
It doesn't cause a problem for me because I'm not cross-compiling
for multiple hosts, but it is annoying.

	$ ./configure
	...
	$ make
	sh: Can't open ../config.guess
	...

Reason: ".." is "$AUX".
The top-level makefile is generated *after* the configure script reruns
itself in the build directory.  At that point, AUX is equal to "..".

Solution: use AM_AUX_DIR_EXPAND to generate an absolute version of
ac_aux_dir.