AX_PROG_DOXYGEN: bugfix
A collegue of mine recently ran into a problem: some "[]" syntax appear
to be leaking through when autoconf generated ./configure. Actually, it
was that a string was over-quoted and was thus not being expanded when
it should have. In some circumstances, it looks like the bug is `mostly
silent' (save for some bogus output in config.log), but we have seen
autoconf fail as a result of the "[]" being taken as an (invalid)
regexp.