GNUmakefile: added make-based build
This build system is nice and simple, but unfortunately, it is slow when
a lot of files have to be re-built, because make invokes a new python
process for every single file that needs to be generated. The scons
build, on the other hand, does all its work within one python process.
That turns out to be significantly faster; apparently, because of high
start-up costs. My guess is that most of those costs are due to
StringTemplate loading (and parsing) its group files.