VL_LIB_READLINE: set LIBS correctly in case of a cache hit
The macro didn't set LIBS appropriately if there is a cache hit. It only
sets it if there is a cache miss.

I've modified the code to set LIBS in the stanza which checks to see if
$vl_cv_lib_readline has been set.  This required resetting LIBS in the
AC_CACHE_CHECK "COMMANDS-TO-SET-IT" argument so that the appropriate
libs wouldn't be included twice if there was a cache miss.