gnulib: deleted submodule
The announce-gen script needs access to the gnulib repository in order
to determine the library's version string. The default location, where
gnulib is expected to be, is $srcdir/gnulib. Quoting from maint.mk:

  gnulib_dir ?= $(srcdir)/gnulib
  gnulib-version = $$(cd $(gnulib_dir) && git describe)

At first, it looked like adding gnulib as a submodule to that location
would be a good idea (especially, since it's needed to bootstrap, too).
After some thought, however, it feels like overkill. Submodules are
meant to track the state of the other repository in the current branch,
but we don't really need that information. We just need some moderately
recent version of gnulib-tool somewhere in $PATH; that's good enough.