AX_LATEX_*: marked obsolete because of issues with LGPL license
The AX_LATEX_* family of macros is distributed under the LPGL license.
Use of that license is discouraged for Autoconf macros, because it might
impose licensing restrictions on the generated configure script. All
other macros in the Archive that are licensed under LGPL or GPL feature
an explicit Autoconf Exception that lifts those limitations, but the
LaTeX macros did not. Unfortunately, all attempts to contact the author
with a request to remedy the situation have been been unsuccessful.