Revert ".prev-version: Delete this file. There was no previous release."
This reverts commit 659842a6f3c9734815c9dcb48781dc96fa3bef76. Gnulib's
release mechanism doesn't work if this file is absent. For the first
release, the generated announcement mail will be wrong, because the
"previous release" denoted in here actually doesn't exist -- so it will
have to be edited appropriately before sending it out --, but after the
first proper release, this file will have the right contents.