set-serial.sh: generate serial numbers automatically
Usage: ./set-serial.sh m4/ax_foo.m4 m4/ax_bar.m4 [...]

This script determines the number of revisions that have occurred to a
given m4 file, and patches an appropriate #serial number into the file
automatically. After that, the script checks whether git(1) regards that
file as modified now. If it does, the serial number is bumped one more
time to ensure that the number will be correct as soon as the changes
have been committed. Consequently, running this script on an unmodified
m4 file with correct serial number information is a no-op and doesn't
change anything.