bootstrap.sh: disable Gnulib's sc_copyright_check target
The regular expression that verifies whether the copyright line in
doc/autoconf-archive.texi is up-to-date assumes that the date is followed by
the string "Free Software Foundation". In our case, however, the copyright
owner is "Autoconf Archive Maintainers". Eventually, the project is going to be
signed over to the FSF; then we can enable that check again.