reorganizing documentation in prep for more prose
On the source side, this splits the documentation into multiple .texi
files and auto-generates the @menu for all-macros.texi, sparing a
redundant edit every time a macro is added or removed.

This also reorganizes the node structure somewhat, using numbered nodes
for everything but the macros, and organizing the macros as sections in
their own chapter, rather than full-fledged chapters.