AX_PREFIX_CONFIG_H: added support for system shell "dash"
The AX_PREFIX_CONFIG_H macro doesn't work with dash, the bug is already
noted in the source file. As of autoconf 2.62, AS_ECHO can be used to
print strings that may contain backslashes or look like options. This
patch updates AX_PREFIX_CONFIG_H to use it.

The comment in the source file refers to $ECHO from libtool, but this
requires libtool.m4 which is not part of autoconf, and AS_ECHO is. On
the other hand, libtool.m4 (at least the one on my system) only requires
autoconf 2.58 or higher, so depending on how much you value
compatibility with older autoconf versions, that may be a better choice.

See <http://savannah.gnu.org/patch/?7317> for further details.