AX_PROG_DOXYGEN: cleaned up use of AM_CONDITIONAL
Previously, calls to AM_CONDITIONAL were inside shell flow control
structures. The automake manual notes that running AM_CONDITIONAL
conditionally will confuse automake. In this case, later use of the
AM_COND_IF to check if Doxygen was found would result in the
conditional appearing to be false, even if it's not.

http://www.gnu.org/s/hello/manual/automake/Usage-of-Conditionals.html