AX_BOOST_SYSTEM, AX_BOOST_FILESYSTEM: inverse the order of search for libraries
With the original macros, the oldest versioned library is used, but
for most distribution this is detrimental, as the old versions are
kept for compatibility, and not to be linked against. By using tac we
reverse the order of the libraries, preferring an unversioned one
(usually the latest available on the system), and then proceeding in
descending order.

Together with this, also try to simplify the sed syntax. It's of note
that even though tac might not be very portable, the sed syntax used
is also not so portable, so it might be worth working this around in a
different way.

Signed-off-by: Diego Elio Pettenò <flameeyes@flameeyes.eu>