Substantial rewrite of AX_CHECK_POSTGRES_DB.
It still behaves the same by default, AC_DEFINing HAVE_LIBPQ and setting
variables PQINCPATH and PQLIBPATH, while erroring when either header or
library is found in the default or specified paths.

The logic is however different, where the specified prefixes and/or
incdir and libdir are temporarily added to CFLAGS and LDFLAGS. The
normal detection logic (AC_CHECK_HEADER and AC_CHECK_LIB) is used next,
only, and once.

In addition, two optional arguments to the macro, ACTION-IF-FOUND and
ACTION-IF-NOT-FOUND are now available to extend the logic. The default
ACTION-IF-FOUND is to do nothing (apart from defining the variables and
substitutions, which is always done). The default ACTION-IF-NOT-FOUND is
to error out; it is replaced by the ACTION-IF-NOT-FOUND if specified
to, e.g., gracefully remove configure PGSQL out when it's only optional.

Bump serial from 5 to 6.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>