AX_CXX_COMPILE_STDCXX_11: test code now includes the "auto" keyword
Furthermore, the macro now uses AC_LANG_PUSH/POP rather than AC_LANG_ASSERT,
which makes it more robust to use.

An optional second argument makes it possible to test for C++11 and return the
result with an AC_DEFINE/AC_SUBST/AM_CONDITIONAL without erroring out if
support is not found.

This new version should be nearly backwards-compatible with the old; the
obvious exception being when compilers which implement a subset of C++11 not
including "auto" are used with codes for which that subset would have been
sufficient.

Further details can be found at <http://savannah.gnu.org/patch/?7985>.