AX_EXT: determine if the OS actually supports AVX
AX_EXT checks for different SIMD variants on a CPU. It checks for AVX, but
doesn't go far enough to actually be useful in some cases. It is possible to be
on a CPU with AVX support and the OS doesn't support it. We should be checking
that such support exists. This is via an additional check of cpuinfo's ECX
register bit 27. Then one needs to call xgetbv and if EAX&0b110 = 0b110 then
AVX is supported by the CPU and OS.

Further details are at <http://savannah.gnu.org/patch/?8084>.