AX_REQUIRE_DEFINED: new helper for requiring macros exist
In cases where we want to require that a macro exists, but not call it,
introduce a new AX_REQUIRE_DEFINED helper.  The existing AC_REQUIRE macro
will call its argument and does not allow for passing of additional args
to the macro.

Once we have this in place, cut a few callers over to it to fix bugs where
they'd accidentally call other helpers w/out any flags.