M4sugar and M4sh added to AX_PROG_JAR, AX_PROG_JAVA, AX_PROG_JAVAC, and AX_PROG_JAVADOC. Also fixed a path discrepancy with AX_JAVA_OPTIONS.
The discrepancy is AX_JAVA_OPTIONS sets up a variable JAVAPREFIX
which is described to the user as the "prefix where the Java runtime
is installed".  Unfortunately, that directory doesn't contain any
executables, but $JAVAPREFIX/bin does.

This incorrect path was passed in the value-if-not-found parameter
of AC_CHECK_PROGS which is incorrect.  I fixed the path and pass it
in to the path parameter.