AX_EXECINFO: new macro to test execinfo size type
<execinfo.h> has been available on a number of systems now, but unfortunately
different systems use different types to represent the size of the backtrace,
with netBSD and FreeBSD using size_t and Linux, Mac OS X. To build calls that
work with both implementations without warning, this macro not only checks if
the execinfo header is available but also what type is used for the len
argument of backtrace_symbols and the return value of backtrace.

All in all this macro defines

    HAVE_EXECINFO_H if execinfo.h can be included,
    backtrace_size_t to the len argument type, i.e. int or size_t,
    HAVE_BACKTRACE to 1 if the function can be linked to,

and append any library needed to LIBS (e.g. some BSD's need -lexecinfo).

Submitted in <https://savannah.gnu.org/patch/index.php?8525>.