AX_LIB_HDF5: remove extra space from include path flag
The code parsing the output of h5cc for the "Installation point:" results in an
extra leading space. The space creeps into HDF5_CPPFLAGS and breaks the include
path flag into two tokens, e.g., "-I" and "/usr/include". While this may be
legal and doesn't affect the compilation, it does break some library
tools (e.g., when the macro is used in building a dynamic library that uses
pkg-config).