cfg.mk: copy 'README.md' to 'README' as part of the 'maintainer-all' target
Github wants the file to be called README.md, but the GNU coding
standards say that the file must be called README. Duh. So we compromise
by keeping the file README.md in the Git repository for Github, but
during the build process we make a suffix-less copy that is the one
we'll also distribute in the release archives.