ax_compiler_flags_gir: Add a compiler flags macro for g-ir-scanner
g-ir-scanner is a tool used within a lot of GNOME projects to generate
introspection data from C APIs. It has a couple of warning flags (and is
not likely to gain any more in the future), and chaining them to
--enable-compile-warnings would be useful.

Integrate AX_COMPILER_FLAGS_GIR into AX_COMPILER_FLAGS so that it’s
enabled by default.