ax_is_release: Add AX_IS_RELEASE to check for compiling release tarballs
This is a macro providing various different policies for determining
whether the code being configured and compiled is from a stable release,
or from a development version. This allows various configuration
settings, such as whether debug is enabled, to be disabled for stable
releases.

A typical example of a policy is to enable stable releases when the .git
directory is not present.