Test that we can actually #include <omp.h> successfully, not just prototype an OpenMP function.
This original version of this configure test actually passes on Apple
LLVM version 6.0 (clang-600.0.51) (based on LLVM 3.5svn), which
doesn't support OpenMP but also doesn't barf when passed -fopenmp.
This can lead to a false positive in codes which need to include the
omp.h header.  The new version of the test addresses this issue.