AX_APPEND_FLAG: fix regression
AX_APPEND_FLAG has changed behaviour compared to earlier versions: it
adds the FLAG also when the variable contains FLAG. I traced down the
altered behaviour to the two lines

 [case " AS_VAR_GET(FLAGS) " in
    *" $1 "*)

which have changed to

  AS_CASE([AS_VAR_GET(FLAGS)],
    ["* $1 *|*$1"]

besides that AS_CASE is used (which is good imho) the pattern has
changed and the space around FLAGS in the first argument of AS_CASE has
gone. This means that e.g. if FLAGS is "alpha beta gamma" that will not
match the pattern "* $1 *|*$1" for $1 being any of 'alpha', 'beta', or
'gamma'. I think that is unexpected behaviour.

This patch restores the original semantics.

For further details can be found at:

  http://lists.gnu.org/archive/html/autoconf-archive-maintainers/2015-02/msg00000.html