ax_compiler_flags: Only check for C++ if AC_PROG_CXX is called
The previous behavior of the macro was to use a C compiler, but when C++
support was added, it created an unconditional dependency on a C++
compiler. This is undesirable for most C projects, and slows down
configure considerably by adding unnecessary checks.

Add an internal _AX_COMPILER_FLAGS_LANG macro, with the desired language
as the first argument. Only require the macro for the corresponding
compiler warning flags if AC_PROG_lang has been called. Enable C support
by default.