ax_compiler_flags_cflags: Remove -Waggregate-return
Historically, this flag has been used because certain old C compilers
didn’t support it — that’s not really relevant any more. It was also
used to warn the developer about the potential for overflowing the
stack. However, I believe the benefits of being able to pass small
structures (e.g. a GdkRectangle) by value outweights the likelihood of
putting a huge structure on the stack and overflowing it.

Within GNOME, a coding standard exists to not return aggregates by value
because it is hard to generate introspected bindings for such code.
g-ir-scanner should warn about this instead.