ax_compiler_flags: Automate addition of -Wno-error=* flags
We need to add a -Wno-error=* flag for each -Wno-* flag as some
compilers will re-enable the warning as an error if -Wall is passed in
the user’s CFLAGS after we’ve specific -Werror.

Previously this was done manually for the -Wno-* flags in
AX_COMPILER_FLAGS_CFLAGS and AX_COMPILER_FLAGS_CXXFLAGS. However, this
ignored the cases where the user was passing -Wno-* flags as extra-flags
to the macros.

Factor this out and automate it so -Wno-error=* flags will be tested and
appended for any -Wno-* flag specified by the user in extra-flags.