ax_compiler_flags: Eliminate ‘minimum’ and ‘maximum’ warning levels
These went against the whole concept of providing a baseline set of
warnings which is set by the module maintainer, since it allowed
individual developers to opt out of certain classes of warning.

Remove them, leaving the ‘no’, ‘yes’ and ‘error’ levels. This maintains
API compatibility of the macros by marking various EXTRA-* variables as
unused, but still handling them, and merging their values with the
preceding EXTRA-* variables. For example, all extra ‘maximum’ and
‘error’ flags are now included in the ‘yes’ level of warnings.