AX_PYTHON_DEVEL: provide PYTHON_LIBS, not PYTHON_LDFLAGS
Because -L and -l are LIBS arguments, not LDFLAGS arguments.
Calling it PYTHON_LDFLAGS suggested wrongly that they should be
used in LDFLAGS, which can cause the -L and -l arguments to
be ignored when they are not in the expected position in the list
of command line arguments. This is a problem, for instance,
with the command line generated by AC_LINK_IFELSE.