AX_PROG_DOXYGEN: provide a DX_RULES substitution
Convert the previously provided aminclude.am snippet in the
documentation into a AC_SUBST substitution. This allows for easy
upgrades to the DX_INIT_DOXYGEN macro without the user having to
manually update the corresponding section of their Makefile.am.

As autoconf substitutions are not recursive, also convert all
@DX_DOCDIR@ and @PACKAGE@ substitutions in the snippet to using
variables. And as AM_CONDITIONAL also relies on substitution,
remove all AM_CONDITIONAL calls and instead use a separate
variable for each of the sections previously enclosed in an if-block.