AX_CODE_COVERAGE: add CPPFLAGS to skip assertions
The assert() statement is a very helpful tool for quality assurance
in nearly every kind of software project. But during code coverage
it unnecessarily decreases the branch coverage percentage as
assertions should never fail. This is probably not what the user
expects as the branch contained inside the assert() statement is not
visible in the code.

So in a workflow where first the "check" target is run to see if any
test fails (e.g. because of a failing assertion) and then the code
coverage is consulted to see how much of the code was exercised, the
assertions should not be counted.

Implement this by adding a variable CODE_COVERAGE_CPPFLAGS which is
defined to -DNDEBUG if CODE_COVERAGE_ENABLED.