AX_PROG_DOXYGEN: allow AC_CONFIG_FILES([Doxyfile])
To allow substitutions in the used Doxyfile, two changes are needed:

1. also substitute everything put in the environment (except DOCDIR)
2. allow Doxyfiles to also reside in the builddir

The latter change is a potentially breaking one! The old behaviour was
to always prefix $(srcdir)/ to the location provided as the second
(and forth, sixth, ...) parameter. This needs to be dropped to allow
for searching in the build directory. To compensate, the default now
is $(srcdir)/Doxyfile.

So the following still works:

DX_INIT_DOXYGEN([foo])
DX_INIT_DOXYGEN([foo], , [dest-dir])

While this needs to be changed:
   DX_INIT_DOXYGEN([foo], [Doxyfile])
-> DX_INIT_DOXYGEN([foo], [$(srcdir)/Doxyfile])