Fix second pass bug in _AX_WITH_CURSES_CHECKEXTRA
Found bug where when _AX_WITH_CURSES_CHECKEXTRA is called a second time
on the same library, e.g. ncurses, the previous setting of
_AX_WITH_CURSES_CHECKEXTRA_cv_var to 'yes' would be changed to 'no' just
due to the tested header file not being available on the second pass.

This was found when AX_WITH_CURSES has 'with_ncurses' set to 'yes' and
'with_ncursesw' set to 'no' forcing _AX_WITH_CURSES_CHECKEXTRA to be
called twice in succession since ax_cv_curses_which is set to 'ncurses'
and the panel.h header file is tested as ncurses/panel.h and panel.h.

The system in question is OpenSuSE LEAP 42.1 which installs
/usr/include/ncurses/panel.h and not /usr/include/panel.h.