Fix ax_boost_system for new versions of boost.
Newer version of boost::system have a breaking change that makes
the test on ax_boost_system fail even though the library is
present.

Changed the test to use another construct that did not change. Also
reseted the "CXX_FLAG" to an empty value in case of it having a
"-Werror" that also makes the test break.