AX_CODE_COVERAGE: add BRANCH_COVERAGE variable
Add a new CODE_COVERAGE_BRANCH_COVERAGE variable that can be used
to provide --rc lcov_branch_coverage=[01] options to lcov and
--rc genhtml_branch_coverage=[01] options to genhtml. These may be
desirable to the user since the default for branch coverage changed
in some "recent" lcov version.

This is achieved by adding a LCOV_SHOPTS variable (for SHared
OPTionS) to the default LCOV_OPTIONS one. And the filtering call of
lcov gets a symmetric LCOV_RMOPTS variable (which is currently
empty).

In addition, ensure that all LCOV_{SHOPTS,OPTIONS,RMOPTS} and
GENHTML_OPTIONS have a corresponding _DEFAULT variable: this became
desirable here because of the new defaults for LCOV_SHOPTS and
GENHTML_OPTIONS; the LCOV_RMOPTS_DEFAULT is there only for
symmetry.