AX_CODE_COVERAGE: fix invalid chars in testname
There was a warning about invalid characters in the testname from
lcov: anything except letters, digits and the underscore causes
this. So replace such characters with underscores (lcov does this
anyway if we don't).

Needs to either call $(subst ) for each undesired character or
needs to execute a system command because GNU make does not
provide regular expressions. Chose here to implement characters
individually. Currently only strips dots and dashes, but these
should be the most common culprits.