AX_VALGRIND_CHECK: clean up configure message
When running configure, the displayed text contained two interleaved
messages. Got:

checking whether to enable Valgrind on the unit tests... checking for valgrind... valgrind
yes

instead of the probably intended:

checking whether to enable Valgrind on the unit tests... yes
checking for valgrind... valgrind

In addition, the first message seems to convey no additional information
compared to the second message: if valgrind is not found (and there is no
--enable-valgrind error), then the result of the check is "no" as
displayed for the executable check. And in all other cases, the result is
"yes". So remove the first message completely: this takes care of the
mangled status messages.