AX_VALGRIND_CHECK: support silent rules
Use the standard autoconf $(V) variable to conditionally hide the
commands executed for check-valgrind: this makes the output from any
test script much more apparent as it is not hidden in clutter.

Hide the outer list of commands in check-valgrind completely with
$(AM_V_at) and define a new $(valgrind_v_use) for the individual
check-valgrind-* calls. The $(valgrind_v_use) call writes:

  USE    $(patsubst check-valgrind-%,%,$@):

Note that this only works inside rules of the form check-valgrind-* .