Add AX_CHECK_OPEN62541_{H,LIB} macros
These macros search for the header and library files for the open62541
library [1]. Support both the amalgamation header and the individual
header file layout. Support both the shared and static library, searched
for in a configurable order. Support for adding the library source
directory and the build directory to the search path if the library is
not installed on the system.

[1]: <http://open62541.org/>