AX_OPEN62541_PATH: fix use of non-inst shared lib
If the open62541 library is not installed and the shared library is to
be used, the --with-open62541=<path> option to contigure setting
CPPFLAGS and LDFLAGS is not enough: other checks by configure may
compile test binaries with the library linked in and then try to execute
then. The execution fails because of the not found shared library object
and gives a wrong test result. Likewise, some make targets (like check)
need to execute compiled binaries.

So modify AX_OPEN62541_PATH() so it exports LD_LIBRARY_PATH with the
absolute path of the build dir for the rest of the configure script to
use and define a Makefile variable OPEN62541_LDPATH, which contains
LD_LIBRARY_PATH=$LD_LIBRARY_PATH:<abs_path_to_build>.