ax_code_coverage: remove lcov version check
The version check does not support the latest version of lcov (1.12).
The history of this module[1][2][3][4] does not reveal a rationale for
the check and its utility seems dubious.  Rather than applying yet another
temporary fix, remove the check.

Note:  The issue of version check removal was raised on
autoconf-archive-maintainers[5] without response.

1.  https://git.gnome.org/browse/gnome-common/log/macros2/gnome-code-coverage.m4
2.  https://bugzilla.gnome.org/show_bug.cgi?id=606720
3.  https://git.gnome.org/browse/glib/log/configure.ac
4.  https://bugzilla.gnome.org/show_bug.cgi?id=501057
5.  https://lists.gnu.org/archive/html/autoconf-archive-maintainers/2016-07/msg00000.html

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>