ax_code_coverage: avoid GNU Make extensions
ifeq/ifneq is a GNU Make extension which doesn't work with BSD Make and
others, resulting in errors such as:

    Missing dependency operator
    Need an operator
    warning: duplicate script for target "ifeq" ignored

Although it would be preferable to use the Automake conditionals
(AM_CONDITIONAL) in the make rules, this does not appear to be possible
within AC_SUBST content.  Instead, build the rules using shell variable
concatenation and only AC_SUBST the ones which should be used.