AX_CHECK_ENABLE_DEBUG: Do not conflict with -DNDEBUG
Since -DNDEBUG is equivalent to "#define NDEBUG 1", explicitly set
the value to ensure that both definitions are equivalent and won't
lead to compiler warnings.

This makes AX_CHECK_ENABLE_DEBUG compatible with AX_CODE_COVERAGE, as
the latter causes -DNDEBUG to be set.

In particular, if AX_COMPILER_FLAGS is also used and thus -Werror
is set (which is the default in that case) this avoids the following
build failure:

./config.h:38:0: error: "NDEBUG" redefined [-Werror]

Submitted in <https://savannah.gnu.org/patch/?9226>

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>