AX_BOOST_BASE: Fix missing BOOST_LDFLAGS.
AX_BOOST_BASE first checks if Boost is installed in a standard location
and sets BOOST_CPPFLAGS and BOOST_LDFLAGS accordingly.

If Boost cannot be found, these variables are reset. Further proceedings
depend on whether the actual location was passed in via
--with-boost=<path> or not. If it was, BOOST_CPPFLAGS is determined
automatically based on the given path. This automatic detection, however, did
not include BOOST_LDFLAGS.

Consequently, if Boost is in a non-standard location that is given via
--with-boost, the user would also have to specify --with-boost-libdir.

This commit fixes this issue by also determining BOOST_LDFLAGS
automatically in this particular branch of execution. The performed
check is very similar to the one done in various other locations of the
macro.