Fix crossbuild for AX_BOOST_BASE
gource fails to cross build from source for wicked reasons. It
terminates with:

| Could not link against -lGLU !

But that is a red herring. GLU is found. It just happens that ax_lib
contains something old. The problem here is that BOOSTLIBDIR ends up
being empty. It is derived from BOOST_LDFLAGS e.g. in AX_BOOST_SYSTEM or
AX_BOOST_FILESYSTEM. Those latter macros expect that BOOST_LDFLAGS
contains a -L flag. That flag goes missing during cross compilation. In
AC_BOOST_BASE, the variable "libsubdirs" is properly set up to contain
the multiarch libdir of the host architecture first. However, it is only
searched for native compilation. For Debian systems, removing the cross
compiling guard would likely resolve the issue. I'm less sure that this
is universally correct though. Maybe a good solution would be to split
"libsubdirs" into the multiarch directory (which would always be
searched) and the others, which would only be searched for native
compilation. I am attaching a patch for this approach.

Signed-off-by: Bastien Roucariès <bastien.roucaries+debian@gmail.com>