ax_cflags_warn_all.m4: Prepend flags instead of appending them
With many compilers (such as the Intel C compilers), flags specified
later in the command line override those specified earlier.
Unfortunately, this means that the user can no longer specify
warning-related options on the command line.  This patch fixes this issue
by using AX_PREPEND_FLAG.

In particular, this patch allows the user to specify "CFLAGS=-w3" on the
./configure command line for the Intel C compiler, which will no longer
be overridden by this macro's "-w2" option.