Unbreak testing for `-std=c++11` with `clang++` with `-Werror -Wnon-virtual-dtor`
Earlier autoconf tests in the [Kyua project](https://github.com/jmmv/kyua)
enable `-Werror -Wnon-virtual-dtor`, which fails because of the fact that
the `Base` and `Derived` classes have non-virtual destructors with
virtual methods. Thus, add a set of virtual destructors for both `Base` and
`Derived` to quell the error and have `clang++` be properly detected as a
C++-11 capable compiler.

Bump serial and add name to copyright for the change.

Tested with:
- clang++ 7.0.1
- g++ 8.2.0

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>