ax_blas: add macOS shared lib globs to with_blas
Prior to this commit, the `with_blas` case statement checked for
dynamic libraries with `.so` suffixes (used on *nix operating
systems), but did not check for dynamic libraries with `.dylib`
suffixes (used on macOS operating systems). This commit adds globs to
the `with_blas` case statement that check for macOS shared libraries
using `.dylib` suffixes.