ax_lapack: add dylib glob to with_lapack
Prior to this commit, the `with_lapack` case statement checked for
dynamic libraries with `.so` suffixes (used on *nix operating
systems), but did not check for dynamic libraries with `.dylib`
suffixes (used on macOS operating systems). This commit adds globs to
the `with_lapack` case statement that check for macOS shared libraries
using `.dylib` suffixes.