BUG FIX: Fixing a _R_CHECK_LENGTH_1_CONDITION_=true bug
If a package exists, the test for its existence produces an error with _R_CHECK_LENGTH_1_CONDITION_=true (introduced in R 3.4.0), e.g.

$ _R_CHECK_LENGTH_1_CONDITION_=true R --silent --vanilla -e 'if(is.na(packageDescription("BH"))) stop("not found")'
> if(is.na(packageDescription("BH"))) stop("not found")
Error in if (is.na(packageDescription("BH"))) stop("not found") :
  the condition has length > 1
Execution halted

This is because 'is.na(packageDescription("BH"))' resolved to a logical vector of 15(!) FALSE values when package 'BH' is installed.  If it is not installed, then it resolves to TRUE.

Bumping 'serial' (because that's what other PRs do)