AX_PTHREAD: use AS_ECHO instead of $as_echo
AS_ECHO(["text"]) is the documented way to perform an 'echo' safely.
$as_echo is an internal, undocumented shell variable that was part of
the implementation of AS_ECHO; it was never intended to be used directly.

The code in ax_pthread.m4 would work with autoconf 2.69, but with
unreleased development versions of Autoconf from January 2013 until
March 2020 (which did not set $as_echo at all) it would have malfunctioned
unpredictably.  With versions after March 2020 it would work correctly
but trigger -Wobsolete warnings.