ax_gcc_func_attribute.m4: test fails for the fallthrough attribute with clang
The test for the fallthrough attribute fails due to other compiler
warnings when using clang.  The following warnings are generated:

   warning: no case matching constant switch condition '0'
   warning: control reaches end of non-void function [-Wreturn-type]

Because of these additional warnings, the test always results in a
failure for the fallthrough attribute.

Update ax_gcc_func_attribute.m4 to eliminate the warnings that are
masking the test for the fallthrough attribute.

Note, tested the entire suite of attribute tests using clang and only
the fallthrough test contained warnings that caused the test to fail.