AX_LIB_HDF5: fix ax_lib_hdf5 on systems with verbose h5cc -show
On some systems, the installation of h5cc is such that `h5cc -show` produces
more than one line. This breaks the macro and produces an invalid Makefile.

This proposed change just calls `head` before the call to `awk` to make sure
only the first line (the relevant one) is used.