Support Debian's posix_local sysconfig scheme in ax_python_devel
Debian adds a custom sysconfig scheme to system python installs,
"posix_local". This is the default scheme, and it redirects local
users' Python module installs to /usr/local even though Python is
installed with a /usr prefix. Both are on Debian's python's sys.path
module search path.

Autoconf and its users understand prefixes, and are likely to select
/usr/local, explicitly. Select the "posix_prefix" scheme, with the
user-supplied prefix.

Previously this custom sysconfig scheme was specified in
distutils.sysconfig, but not sysconfig itself. As distutils is being
deprecated, the custom scheme is now specified in sysconfig, since
Debian's Python 3.10 (3.10.2-4).